+++ src/main/kotlin/pythagoras/f/Ellipse.kt
@@ -24,16 +24,16 @@ package pythagoras.f
 class Ellipse : AbstractEllipse {
 
     /** The x-coordinate of the framing rectangle.  */
-    override var x: Float = 0.toFloat()
+    override var x: Float = 0f
 
     /** The y-coordinate of the framing rectangle.  */
-    override var y: Float = 0.toFloat()
+    override var y: Float = 0f
 
     /** The width of the framing rectangle.  */
-    override var width: Float = 0.toFloat()
+    override var width: Float = 0f
 
     /** The height of the framing rectangle.  */
-    override var height: Float = 0.toFloat()
+    override var height: Float = 0f
 
     /**
      * Creates an ellipse with framing rectangle (0x0+0+0).
