+++ src/main/kotlin/pythagoras/f/Matrix3.kt
@@ -25,15 +25,15 @@ import java.lang.Math
  * The property names take the form `mCOLROW`.
  */
 data class Matrix3(
-        override var m00: Float = 0.toFloat(),
-        override var m10: Float = 0.toFloat(),
-        override var m20: Float = 0.toFloat(),
-        override var m01: Float = 0.toFloat(),
-        override var m11: Float = 0.toFloat(),
-        override var m21: Float = 0.toFloat(),
-        override var m02: Float = 0.toFloat(),
-        override var m12: Float = 0.toFloat(),
-        override var m22: Float = 0.toFloat()
+        override var m00: Float = 0f,
+        override var m10: Float = 0f,
+        override var m20: Float = 0f,
+        override var m01: Float = 0f,
+        override var m11: Float = 0f,
+        override var m21: Float = 0f,
+        override var m02: Float = 0f,
+        override var m12: Float = 0f,
+        override var m22: Float = 0f
 ) : IMatrix3 {
     /**
      * Creates an identity matrix.
